- Check if updates are needed prior to doing anything - Try to download binary 3 times before continuing - Deletion of previous files is now granular

+++ install.sh
@@ -1,34 +1,57 @@
 #!/bin/bash
 # Author: CryoByte33
 
+# Create a hidden directory for the script, if not present
+mkdir -p "$HOME/.cryo_utilities" &>/dev/null
+cd "$HOME/.cryo_utilities" || exit 1
+
+# Download checksum to compare with local binary, if present
+wget https://github.com/CryoByte33/steam-deck-utilities/releases/download/latest/cu.md5 -O "$HOME/.cryo_utilities/cu.md5" 2>&1
+if md5sum -c --quiet cu.md5; then
+  zenity --info --text="No update necessary!" --width=300
+  exit 0
+fi
+
 # Uninstall swap resizer if present
-# Delete old install directory
+# Delete legacy install directory
 rm -rf "$HOME/.swap_resizer" &>/dev/null
 
-# Remove old Desktop icons
+# Remove legacy Desktop icons
 rm -rf ~/Desktop/SwapResizerUninstall.desktop &>/dev/null
 rm -rf ~/Desktop/SwapResizer.desktop &>/dev/null
 
-# Remove directory if present
-rm -rf "$HOME/.cryo_utilities" &>/dev/null
+# Remove old files
+rm -f "$HOME/.cryo_utilities/cryo_utilities" &>/dev/null
+rm -f "$HOME/.cryo_utilities/launcher.sh" &>/dev/null
+rm -f "$HOME/.cryo_utilities/cryo-utilities.png" &>/dev/null
 
-# Create a hidden directory for the script
-mkdir -p "$HOME/.cryo_utilities" &>/dev/null
+# Attempt to download the binary 3 times.
+for i in {1..3}; do
+  # Download binary
+  wget https://github.com/CryoByte33/steam-deck-utilities/releases/download/latest/cryo_utilities -O "$HOME/.cryo_utilities/cryo_utilities" 2>&1 | sed -u 's/.* \([0-9]\+%\)\ \+\([0-9.]\+.\) \(.*\)/\1\n# Downloading at \2\/s, ETA \3/' | zenity --progress --title="Downloading CU Binary, attempt $i of 3..." --auto-close --width=500
 
-# Install binary
-wget https://github.com/CryoByte33/steam-deck-utilities/releases/download/latest/cryo_utilities -O "$HOME/.cryo_utilities/cryo_utilities" 2>&1 | sed -u 's/.* \([0-9]\+%\)\ \+\([0-9.]\+.\) \(.*\)/\1\n# Downloading at \2\/s, ETA \3/' | zenity --progress --title="Downloading CU Binary" --auto-close --width=300
-
-# Start a loop testing if zenity is running, and if not kill wget (allows for cancel to work)
-RUNNING=0
-while [ $RUNNING -eq 0 ]; do
-  if [ -z "$(pidof zenity)" ]; then
-    pkill wget
-    RUNNING=1
+  # Start a loop testing if zenity is running, and if not kill wget (allows for cancel to work)
+  RUNNING=0
+  while [ $RUNNING -eq 0 ]; do
+    if [ -z "$(pidof zenity)" ]; then
+      pkill wget
+      RUNNING=1
+    fi
+    sleep 0.1
+  done
+  # Compare checksum to new binary
+  if md5sum -c --quiet cu.md5; then
+    break
   fi
-  sleep 0.1
+
+  if [ "$i" -ge "3" ]; then
+    zenity --error --text="Install/upgrade of CryoUtilities has failed!\n\nBinary couldn't be downloaded correctly, this may be a network or GitHub issue." --width=500
+    exit 1
+  fi
 done
 
 chmod +x "$HOME/.cryo_utilities/cryo_utilities"
+rm -f cu.md5 &>/dev/null
 
 # Install launcher script
 wget https://raw.githubusercontent.com/CryoByte33/steam-deck-utilities/main/launcher.sh -O "$HOME/.cryo_utilities/launcher.sh"
@@ -36,7 +59,6 @@ chmod +x "$HOME/.cryo_utilities/launcher.sh"
 
 # Install Icon
 wget https://raw.githubusercontent.com/CryoByte33/steam-deck-utilities/main/cmd/cryoutilities/Icon.png -O "$HOME/.cryo_utilities/cryo-utilities.png"
-cd ~/.cryo_utilities || exit
 xdg-icon-resource install cryo-utilities.png --size 64
 
 # Create Desktop icons
